requirements.txt
fastapi
uvicorn
python-telegram-bot
mysql-connector-python